Remove kinetic energy from SystemState
We don't want to compute the kinetic energy every time step, and we
don't want to leave the SystemState in an invalid state after any time
steps.  Since the kinetic energy is just a derived quantity that can be
computed from the velocities, we will put its computation elsewhere and
not store it in the SystemState.