Add operator() overload for multiple time steps
Integrator was only written to advance the SystemState by one time step,
but it would be useful to give it a specific number of time steps to
advance, so that we do not have to write loops everywhere.  Added this
functionality and a clause in the unit tests for it.