Add center argument to angular momentum, inertia
It is useful to compute the angular momentum and inertia tensor about
any given center.  While these could in principle be obtained from the
quantities about zero using the parallel axis theorem, it is most
convenient to simply define a function overload that accepts the center
as an argument and uses it.