Add initial_condition.hpp, .cpp
Wrote the InitialCondition class which accepts data which are
physically meaningful to the user (number of particles, density,
temperature, random seed), and uses them to build the appropriate
initial state information for the simulation.

We must construct both a BoundingBox and a SystemState.  We use a chain
of delegating constructors in order to do this in the initializer list,
which allows us to avoid dynamic allocation of this objects.  The
InitialState object then keeps a copy of the BoundingBox and
SystemState, as well as the other information provided (such as the
random seed), which allows the simulation to be easily repeatable.