Add cubic_lattice
The InitialCondition code is kind of ugly and I think the best way to
refactor it is to add a CubicLattice tool that handles the details of
generating the lattice sites.  Then InitialCondition should only need
to be concerned with the physically relevant quantities that are given
in its constructor.

We still need to change InitialCondition to actually use the new
CubicLattice tool, as well as write unit tests.