Fix scale factor calculation
Originally we were computing the density of the prototype crystal
lattice with the assumption that it was the same as the density of an
individual unit cell.  However, if the cubical box is underfilled, this
is no longer true.  We have fixed this calculation to use the total
particle count and the total volume in a straightforward way.

Also, we move these computations to the body of the constructor for
clarity.