Add test for CubicLattice
Added test cases for CubicLattice.  Second test case is commented out,
because strange things happen if it is left in place.

I have learned that the C++20 standard does not properly define how the
coroutine state should be allocated, when a coroutine needs to be
suspended and resumed.  The result is that it is possible for Eigen
fixed-size types to be improperly aligned.  This never occurred in my
previous unit testing, but for some reason these tests with CubicLattice
cause the problem to show up.  Will probably have to do some redesign
in order to avoid this misalignment.