Fix alignment of coroutine frame for Eigen
Was running into problems with some unit tests, and tracked them down
to a bug in GCC (https://gcc.gnu.org/bugzilla/show_bug.cgi?id=104177).
For whatever reason, GCC fails to apply overalignment for stack-
allocated local variables when allocating the coroutine frame.  During
the course of investigating this issue, I also learned that the standard
itself has a defect in its description of how coroutine frames should
be allocated.  There is ongoing discussion on this issue, so it may be
a while before it is fixed.

In the meantime, I have changed some of the code in generator.hpp
to allow manual alignment of the coroutine frame.  To help support this
change, I have added aligned_generator.hpp which simply defines some
useful aliases to cut down on the verbosity required.

All files using std::generator had to be updated.  After much testing,
everything appears to work properly now.