Refactor InitialCondition to use CubicLattice
The reason for writing CubicLattice was because InitialCondition was
getting unwieldy, and especially the code used to construct it and set
up the initial particle positions was quite convoluted and beginning
to feel like it was taking on too much responsibility.  CubicLattice
allows us to delegate most of that responsibility, so that
InitialCondition construction can be simplified as well as focus mainly
on physically-relevant notions.