Add test for equilibrator
Added a test to verify that Equilibrator works.  It turns out this was
tricky to set up and required remembering some important physics:
namely, that thermodynamic temperature is NOT instantaneous temperature.
Temperature is proportional to the average kinetic energy, but an
individual time step may be in a very atypical state.  So it is
necessary to average over several time steps.