Fix type safety issues
The way moving_sample was previously written, it could have possibly
attempted to return the wrong types if asked to compute the mean and
variance of a sample of integers.  This is not an actual use case,
but to make the code more robust, I have modified the templates so that
integer samples receive integer division.  This will simply produce
nonsense results, instead of failing to compile.

I have no plans to actually make the computation "correct" for samples
of integers, as this use case does not make sense.