Change State-modifying methods to Operators
SystemState previously contained a method clear_dynamical_quantities
which felt a bit strange alongside the Operator notion for all other
functions which modify the state.  This method was moved outside the
SystemState class and is now an Operator.  We also added
clear_displacements, to be used in a similar manner.