Add Commands and SimulationPhase
We will use a Command Pattern to allow the SimulationPhase to control
the flow of operations during the Simulation while not requiring
SimulationPhase to _depend_ on Simulation.  This will allow Simulation
to run the main loop, and SimulationPhase to merely encapsulate the
decision-making processes that goven the Equilibration and Observation
phases.