Define Measurement concept and some examples
Decided it would be useful to have Measurement objects that use the
same piping syntax as Operators, except that they do not change the
SystemState they act on.  This will be useful for collecting several
measurements into one object.  The only downside is that these
measurements need to be stored as a side effect.