Define Measurements and Properties
Defined two further Concepts that use the piping syntax:

    state | operator | measurement | property;

To summarize:

1. Operators modify the state and return a reference for further piping.
2. Measurements do not modify the state, and return a const reference
for further piping.  They are expected to use side effects to record
whatever information they are looking for.
3. Properties accept a const reference and return a value (so that
piping cannot continue after a Property is encountered).

The intent is to provide a uniform syntax for accessing the SystemState,
so that it stands out in the code.