Add simulation_phase.hpp
Defined the mechanism for implementing "phases" of the simulation while
separating concerns as much as possible.  The simulation will be driven
via a global loop, but a SimulationPhase object will make the actual
decisions about what to do each loop iteration, making use of a Command
pattern.

This way each SimulationPhase doesn't have to drive its own loop, and
there can be a global notion of time.  The Command pattern allows us to
avoid the SimulationPhase having a circular dependency on the Simulation
loop itself.  The cost of this is that the Simulation loop will have to
take responsibility for implementing the Commands that have been
defined.