Move parameter structs outside classes
Originally the EquilibrationParameters and ObservationParameters were
nested structs inside EquilibrationPhase and ObservationPhase.
Logically, this makes sense; however there is a GCC bug which prevents
us from being able to use the default values in the EquilibrationPhase
and ObservationPhase constructors.  Moving them outside so that they are
un-nested resolves this problem.