Add Analyzers to compute time-averaged quantities
Refactored the approach for computing time-averaged physical quantities
like temperature, pressure, etc.  Removed observation.hpp and replaced
with analyzers.hpp.  Defined Analyzer classes both for simple
temperature computation as well as for the more complex computation of
all the quantities in an Observation.  This way the two computations
are unified under the same interface, and also all computation of
physical quantities remains in the physics namespace/library.