Move Parameters structs back inside Phases
It really makes more sense to have the Parameters structs inside the
SimulationPhase class definitions as nested classes.  To get around the
GCC bug described in (https://gcc.gnu.org/bugzilla/show_bug.cgi?id=88165),
we simply declare a second constructor overload which is defined in the
cpp file.  At this point in the translation unit, the outer class is
fully defined, so the bug does not occur and we can call the nested
class's default constructor without problem.