Add set_start_time() method
It is much simpler for SimulationPhase to have a set_start_time()
method than to have to pass this value in the constructor (and then
have to ensure that all SimulationPhases get constructed at the time
that this value is known).  With a method to explicitly set the start
time, we can simply pass a list of SimulationPhases as the "schedule"
and initialize them at the time they are needed.