Add AdvanceTime command
Added a "default" command to instruct the main Simulation to use the
Integrator to advance time on the SystemState.  This provides a uniform
interface and greatly simplifies the main simulation loop.

At the moment, the SimulationPhase is directly responsible for issuing
the AdvanceTime command to the Simulation so that it will proceed to
the next time step, otherwise the main simulation loop exits.  I need
to consider whether this makes the AbortSimulation and PhaseComplete
commands partially redundant, since now the loop ends unless explicitly
commanded to continue.  Since there are multiple reasons the loop can
end, at the moment I think the current arrangement makes sense
(especially considering that logging functions still need to be added,
which will record various events that happen such as starting/ending
SimulationPhases and/or aborting the simulation).