Add sinks, log_message, clean up many files
Taking a step back to implement the lower-level IO functionality before
changing the Simulation too much.  We have defined a std::variant type
for log messages, and then a collection of Sink classes which actually
format the messages into strings and print them to files.

With these tools, we can begin to build the file IO structure that will
run in a parallel thread to the main simulation.