Modify test_sinks to use actual files
Changed the test case so that it writes to actual files and reads them
back, rather than simply writing to a std::stringstream.

Mostly wanted to understand how to control where files get written
when using CTest/Catch2.  Also learned a bit about how to use
std::filesystem, including how to write the test files to the standard
temp directory.

Amended: We do not write to the system temp directory, because we
cannot guarantee unique names.  There is probably a way to do this with
Boost.