Change to unique pointers
The various polymorphic classes which compose the parts of the
simulation need to be heap-allocated at runtime.  The previous code
structure used member variables which were references (or sometimes
const references), which assume the existence of an instance whose
lifetime is managed externally.  These start to look like global
variables, and they complicate the effort to make a factory function
for creating Integrators, etc.  So, we change these to std::unique_ptr
so that they can be dynamically allocated, but with lifetimes in the
local scope.

Have changed the member variables to unique pointers in:

1. integrator.hpp
2. force_calculation.hpp

This required many adjustments to other files due to changing the
interfaces.

Also added virtual destructors on all classes that are dynamically
allocated.