Implement Step Builder pattern
This is a much more convenient way to build various Integrator types,
since the various options you have at each step are dependent on the
options you've chosen in previous steps.  Not sure if this is the best
implementation or not, but it uses a mix of function arguments and
template parameters to guide the building process, with sensible
defaults.

For the interface with Python, it will have to be wrapped by some
factory function that takes string/dictionary inputs, but this will
be quite complex.  The various paths in the "options tree" will need to
be given as something like a JSON string.