Change ownership of ShortRangeForce
Rather than pass a unique_ptr which transfers ownership to the
ForceCalculation (and ultimiately, to the SimulationController),
I've changed back to using a const & so that ownership is retained by
the caller.  This way, the ShortRangeForce can be reused in other ways,
such as allowing the user to create a plot of the potential used in
the simulation.