Define ShortRangeForce::Parameters struct
Changed every ShortRangeForce to use a Parameters struct.  Although the
existing ShortRangeForce types only have two constructor arguments, the
Parameters struct gives us a uniform interface and also makes it easier
to construct a configuration object that can contain both the type of
ShortRangeForce needed as well as the Parameters used to construct it.