Experiment with console output to Python
Using the asynchronous IO architecture that we have currently set up,
the output from Simulation::run() does not show up "live" in a Jupyter
notebook, but instead shows up all at once at the end of the process.

This is not what we want.  I have created a dummy Worker class to test
some ideas for how to get it to work properly.  The correct architecture
is the one that is now set up in this commit: the Worker launches some
job asynchronously (in this case, in a detached thread), and this job
sends messages back to the main thread via a message queue.  Then we can
forward strings to Python.

In order for live, incremental output to appear in a Jupyter notebook,
it _must_ come from a Python print() function, and not from anything on
the C++ side.  Therefore Python itself must become the consumer process
of the producer/consumer buffer.  So, as shown here, we must re-
architect things so that the _simulation_ runs in a background thread,
and Python is ultimately responsible for printing to the screen in the
main thread.  We may also decide to let Python do file IO as well.