Add paths.py helper functions for unit tests
For C++ unit testing, it was simple to just specify the working
directory in which to run the tests, since they are compiled artifacts
anyway.  In Python, this is not so nice, since the Python unit tests
themselves are just a bunch of .py files in directories.  There are
options for Pytest related to work directories, but in my limited
experimentation with them, it seemed quite easy to break imports if
the working directory was set to something outside the tests/ directory.

So I've just collected a few useful things in paths.py for finding the
temp directory in which to create files for reading back.