Add simulation_result.py
Added an enum type to track the "result" of a simulation (whether it
completed or aborted, and if it aborted, for what reason), and a simple
function for obtaining this information from the events log.

This will mainly be useful for simulation sweeps, where we'd like to
run a whole collection of simulations and then understand immediately
which ones may have failed.